+++ chroot/run_linux.go
@@ -4,11 +4,9 @@
 package chroot
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -17,14 +15,11 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"syscall"
 	"time"
 
-	"github.com/containers/buildah/bind"
 	"github.com/containers/buildah/copier"
 	"github.com/containers/storage/pkg/mount"
-	"github.com/containers/storage/pkg/unshare"
 	"github.com/opencontainers/runc/libcontainer/apparmor"
 	"github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/sirupsen/logrus"
@@ -59,145 +54,6 @@ type runUsingChrootSubprocOptions struct {
 	BundlePath  string
 	UIDMappings []syscall.SysProcIDMap
 	GIDMappings []syscall.SysProcIDMap
-}
-
-// runUsingChroot, still in the grandparent process, sets up various bind
-// mounts and then runs the parent process in its own user namespace with the
-// necessary ID mappings.
-func runUsingChroot(spec *specs.Spec, bundlePath string, ctty *os.File, stdin io.Reader, stdout, stderr io.Writer, closeOnceRunning []*os.File) (wstatus unix.WaitStatus, err error) {
-	var confwg sync.WaitGroup
-
-	// Create a new mount namespace for ourselves and bind mount everything to a new location.
-	undoIntermediates, err := bind.SetupIntermediateMountNamespace(spec, bundlePath)
-	if err != nil {
-		return 1, err
-	}
-	defer func() {
-		if undoErr := undoIntermediates(); undoErr != nil {
-			logrus.Debugf("error cleaning up intermediate mount NS: %v", err)
-		}
-	}()
-
-	// Bind mount in our filesystems.
-	undoChroots, err := setupChrootBindMounts(spec, bundlePath)
-	if err != nil {
-		return 1, err
-	}
-	defer func() {
-		if undoErr := undoChroots(); undoErr != nil {
-			logrus.Debugf("error cleaning up intermediate chroot bind mounts: %v", err)
-		}
-	}()
-
-	// Create a pipe for passing configuration down to the next process.
-	preader, pwriter, err := os.Pipe()
-	if err != nil {
-		return 1, fmt.Errorf("error creating configuration pipe: %w", err)
-	}
-	config, conferr := json.Marshal(runUsingChrootExecSubprocOptions{
-		Spec:       spec,
-		BundlePath: bundlePath,
-	})
-	if conferr != nil {
-		fmt.Fprintf(os.Stderr, "error re-encoding configuration for %q", runUsingChrootExecCommand)
-		os.Exit(1)
-	}
-
-	// Apologize for the namespace configuration that we're about to ignore.
-	logNamespaceDiagnostics(spec)
-
-	// If we have configured ID mappings, set them here so that they can apply to the child.
-	hostUidmap, hostGidmap, err := unshare.GetHostIDMappings("")
-	if err != nil {
-		return 1, err
-	}
-	uidmap, gidmap := spec.Linux.UIDMappings, spec.Linux.GIDMappings
-	if len(uidmap) == 0 {
-		// No UID mappings are configured for the container.  Borrow our parent's mappings.
-		uidmap = append([]specs.LinuxIDMapping{}, hostUidmap...)
-		for i := range uidmap {
-			uidmap[i].HostID = uidmap[i].ContainerID
-		}
-	}
-	if len(gidmap) == 0 {
-		// No GID mappings are configured for the container.  Borrow our parent's mappings.
-		gidmap = append([]specs.LinuxIDMapping{}, hostGidmap...)
-		for i := range gidmap {
-			gidmap[i].HostID = gidmap[i].ContainerID
-		}
-	}
-
-	// Start the parent subprocess.
-	cmd := unshare.Command(append([]string{runUsingChrootExecCommand}, spec.Process.Args...)...)
-	setPdeathsig(cmd.Cmd)
-	cmd.Stdin, cmd.Stdout, cmd.Stderr = stdin, stdout, stderr
-	cmd.Dir = "/"
-	cmd.Env = []string{fmt.Sprintf("LOGLEVEL=%d", logrus.GetLevel())}
-	cmd.UnshareFlags = syscall.CLONE_NEWUTS | syscall.CLONE_NEWNS
-	requestedUserNS := false
-	for _, ns := range spec.Linux.Namespaces {
-		if ns.Type == specs.UserNamespace {
-			requestedUserNS = true
-		}
-	}
-	if len(spec.Linux.UIDMappings) > 0 || len(spec.Linux.GIDMappings) > 0 || requestedUserNS {
-		cmd.UnshareFlags = cmd.UnshareFlags | syscall.CLONE_NEWUSER
-		cmd.UidMappings = uidmap
-		cmd.GidMappings = gidmap
-		cmd.GidMappingsEnableSetgroups = true
-	}
-	if ctty != nil {
-		cmd.Setsid = true
-		cmd.Ctty = ctty
-	}
-	cmd.OOMScoreAdj = spec.Process.OOMScoreAdj
-	cmd.ExtraFiles = append([]*os.File{preader}, cmd.ExtraFiles...)
-	interrupted := make(chan os.Signal, 100)
-	cmd.Hook = func(int) error {
-		for _, f := range closeOnceRunning {
-			f.Close()
-		}
-		signal.Notify(interrupted, syscall.SIGHUP, syscall.SIGINT, syscall.SIGTERM)
-		go func() {
-			for receivedSignal := range interrupted {
-				if err := cmd.Process.Signal(receivedSignal); err != nil {
-					logrus.Infof("%v while attempting to forward %v to child process", err, receivedSignal)
-				}
-			}
-		}()
-		return nil
-	}
-
-	logrus.Debugf("Running %#v in %#v", cmd.Cmd, cmd)
-	confwg.Add(1)
-	go func() {
-		_, conferr = io.Copy(pwriter, bytes.NewReader(config))
-		pwriter.Close()
-		confwg.Done()
-	}()
-	err = cmd.Run()
-	confwg.Wait()
-	signal.Stop(interrupted)
-	close(interrupted)
-	if err != nil {
-		if exitError, ok := err.(*exec.ExitError); ok {
-			if waitStatus, ok := exitError.ProcessState.Sys().(syscall.WaitStatus); ok {
-				if waitStatus.Exited() {
-					if waitStatus.ExitStatus() != 0 {
-						fmt.Fprintf(os.Stderr, "subprocess exited with status %d\n", waitStatus.ExitStatus())
-					}
-					os.Exit(waitStatus.ExitStatus())
-				} else if waitStatus.Signaled() {
-					fmt.Fprintf(os.Stderr, "subprocess exited on %s\n", waitStatus.Signal())
-					os.Exit(1)
-				}
-			}
-		}
-		fmt.Fprintf(os.Stderr, "process exited with error: %v", err)
-		os.Exit(1)
-	}
-
-	return 0, nil
 }
 
 // main() for parent subprocess.  Its main job is to try to make our
